Update to Central Portal publishing with gradle-maven-publish-plugin
The pom-default.xml is exactly the same.

The module.json is _almost_ the same, with the one difference being the removal of the javadocElements variant. It is unclear if this has any material difference. (The javadoc jar is still published.) Noted in https://github.com/vanniktech/gradle-maven-publish-plugin/issues/1028.

See https://vanniktech.github.io/gradle-maven-publish-plugin/